Correct use of TestDirectory
The continuous integration setup requires that test failures are signaled by
the exit code of the GAP test scripts (i.e., tst/testall.g). This is what
`TestDirectory` normally ensures. But by calling it several times, only
allowing the final call to set a non-zero exit code, the results of all but
the final `TestDirectory` call are discarded.

To fix this, simply call `TestDirectory` once but with a list of all
directories to be processed.