Tweak how CrossedInvariant uses AllHomomorphisms
Instead of invoking AllHomomorphisms on an fp-group, invoke it
on an isomorphic permutation group. This is faster. It also avoids
a problem with GAP 4.12 where AllHomomorphisms runs into an error
when invoked on a solvable fp group (that bug will be fixed there,
but in the meantime, it seems sensible to improve the HAP code, too)