Again: Don't call ObjectifyWithAttributes on a string
Objectify and ObjectifyWithAttributes must only be called on records or
plain lists. Unfortunately this was not enforced uniformly by the GAP
kernel, but in future versions it may be.

A similar patch was already merged in January 2020, but it was reverted
(probably by accident) in April 2020 (in a commit with titled "Add
directory Congruence, plus related files", hence why I think this was an
accident).